# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Remote Controller Store API
  version: "0.1.0"
  description: HTTP API for the remote Juju controller store daemon (rcsd).
paths:
...
  /controllers/{name}/models:
    parameters:
      - $ref: '#/components/parameters/ControllerName'
    get:
      operationId: AllModels
      summary: List all models for a controller
      responses:
        "200":
          description: Map of model name to details JSON
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
                  additionalProperties: true
        "404":
          description: No models found
          content:
            text/plain:
              schema:
                type: string
    put:
      operationId: SetModels
      summary: Replace all models for a controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: object
                additionalProperties: true
      responses:
        "204":
          description: Replaced
  /controllers/{name}/models/current:
    parameters:
      - $ref: '#/components/parameters/ControllerName'
    get:
      operationId: CurrentModel
      summary: Get the current model for a controller
      responses:
        "200":
          description: Current model name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NameResponse'
        "404":
          description: No current model
          content:
            text/plain:
              schema:
                type: string
    put:
      operationId: SetCurrentModel
      summary: Set the current model for a controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
      responses:
        "204":
          description: Updated
        "404":
          description: Model not found
          content:
            text/plain:
              schema:
                type: string
  /controllers/{name}/models/previous:
    parameters:
      - $ref: '#/components/parameters/ControllerName'
    get:
      operationId: PreviousModel
      summary: Get the previous model for a controller
      responses:
        "200":
          description: Previous model name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NameResponse'
        "404":
          description: No previous model
          content:
            text/plain:
              schema:
                type: string
  /controllers/{name}/models/{model}:
    parameters:
      - $ref: '#/components/parameters/ControllerName'
      - $ref: '#/components/parameters/ModelName'
    get:
      operationId: ModelByName
      summary: Get a specific model by name
      responses:
        "200":
          description: Model details
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: Not found
          content:
            text/plain:
              schema:
                type: string
    put:
      operationId: UpdateModel
      summary: Create or update a model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "204":
          description: Updated
    delete:
      operationId: RemoveModel
      summary: Remove a model
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
          content:
            text/plain:
              schema:
                type: string
...
components:
  parameters:
    ControllerName:
      name: name
      in: path
      required: true
      schema:
        type: string
    ModelName:
      name: model
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    NameRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    NameResponse:
      type: object
      properties:
        name:
          type: string
...